Unbelievably, this works well
