+++ controller/src/TrainingWheels/Plugin/GitFiles/GitFiles.php
@@ -9,4 +9,11 @@ class GitFiles extends PluginBase {
     parent::__construct();
     $this->ansible_play = __DIR__ . '/ansible/gitfiles.yml';
   }
+
+  public function mixinEnvironment($env, $type) {
+    if ($type == 'linux') {
+      $gitFilesLinuxEnv = new GitFilesLinuxEnv();
+      $gitFilesLinuxEnv->mixinLinuxEnv($env);
+    }
+  }
 }
